+++ include/internal/cryptlib.h
@@ -258,4 +258,6 @@ static ossl_inline int ossl_is_absolute_path(const char *path)
     return path[0] == '/';
 }
 
+int ossl_init_casecmp_int(void);
+void ossl_deinit_casecmp(void);
 #endif
